OptionalWorkloadProvider rename and split of files
